+++ packages/workspace-fusion/src/lib/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useResizeObserver';
+export * from './useWorkspaceController';
